review: move to `{iroha_executor::default, iroha_executor_data_model}::custom::multisig`
