+++ crates/iroha/Cargo.toml
@@ -52,12 +52,12 @@ iroha_config = { workspace = true }
 iroha_config_base = { workspace = true }
 iroha_crypto = { workspace = true }
 iroha_data_model = { workspace = true, features = ["http"] }
+iroha_executor_data_model = { workspace = true }
 iroha_primitives = { workspace = true }
 iroha_logger = { workspace = true }
 iroha_telemetry = { workspace = true }
 iroha_torii_const = { workspace = true }
 iroha_version = { workspace = true }
-iroha_multisig_data_model = { workspace = true }
 
 attohttpc = { version = "0.28.0", default-features = false }
 eyre = { workspace = true }
@@ -84,7 +84,6 @@ nonzero_ext = { workspace = true }
 iroha_genesis = { workspace = true }
 iroha_test_samples = { workspace = true }
 iroha_test_network = { workspace = true }
-iroha_executor_data_model = { workspace = true }
 executor_custom_data_model = { version = "=2.0.0-rc.1.0", path = "../../wasm/samples/executor_custom_data_model" }
 
 tokio = { workspace = true, features = ["rt-multi-thread"] }
