+++ crates/iroha/src/lib.rs
@@ -9,4 +9,4 @@ mod secrecy;
 
 pub use iroha_crypto as crypto;
 pub use iroha_data_model as data_model;
-pub use iroha_multisig_data_model as multisig_data_model;
+pub use iroha_executor_data_model as executor_data_model;
