+++ crates/iroha/tests/multisig.rs
@@ -9,7 +9,7 @@ use iroha::{
     client::Client,
     crypto::KeyPair,
     data_model::{prelude::*, Level},
-    multisig_data_model::*,
+    executor_data_model::custom::multisig::*,
 };
 use iroha_test_network::*;
 use iroha_test_samples::{
