+++ crates/iroha_executor/Cargo.toml
@@ -17,7 +17,6 @@ debug = ["iroha_smart_contract/debug"]
 iroha_executor_derive = { path = "../iroha_executor_derive" }
 
 iroha_executor_data_model.workspace = true
-iroha_multisig_data_model.workspace = true
 iroha_smart_contract_utils.workspace = true
 iroha_smart_contract.workspace = true
 iroha_schema.workspace = true
