+++ crates/iroha_executor_data_model/Cargo.toml
@@ -16,5 +16,7 @@ iroha_executor_data_model_derive = { path = "../iroha_executor_data_model_derive
 iroha_data_model.workspace = true
 iroha_schema.workspace = true
 
+derive_more = { workspace = true, features = ["constructor", "from"] }
+parity-scale-codec.workspace = true
 serde.workspace = true
 serde_json.workspace = true
