+++ wasm/libs/default_executor/Cargo.toml
@@ -11,7 +11,6 @@ crate-type = ['cdylib']
 
 [dependencies]
 iroha_executor.workspace = true
-iroha_multisig_data_model.workspace = true
 
 panic-halt.workspace = true
 dlmalloc.workspace = true
